requirements.txt
Flask
fpdf
requests
pydub
google-generativeai